refactor: 헤더 여부에 따라 outlet 크기 조절 (#35)
