+++ src/components/layout/Layout.tsx
@@ -2,14 +2,14 @@ import * as S from './Layout.styles';
 import { Outlet } from 'react-router-dom';
 import Header from './Header/Header';
 import Nav from './Nav/Nav';
-import { useAuth } from '@/hooks';
+import { useShowHeader } from '@/hooks';
 
 export const Layout = () => {
-  const { user } = useAuth();
+  const showHeader = useShowHeader();
   return (
     <S.MainContainer>
-      <Header />
-      <S.OutletContainer id="outlet" $hasHeader={!!user}>
+      <Header showHeader={showHeader} />
+      <S.OutletContainer id="outlet" $showHeader={showHeader}>
         <Outlet />
       </S.OutletContainer>
       <Nav />
