+++ src/hooks/index.ts
@@ -1,5 +1,5 @@
 export * from './queries';
 export * from './auth';
+export * from './ui';
 
 export { default as useErrorHandler } from './useErrorHandler';
-export { default as useAlert } from './useAlert';
